10/05/2023

+++ mycontacts-backend/controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async(req,res) => {
-    const contacts = await Contact.find()
+    const contacts = await Contact.find({user_id: req.user.id})
     res.status(200).json({contacts});
 });
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async(req,res) => {
     const contact = await Contact.findById(req.params.id);
@@ -25,7 +25,7 @@ const getContact = asyncHandler(async(req,res) => {
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async(req,res) => {
     const{name, email, phone} = req.body;
@@ -35,14 +35,14 @@ const createContact = asyncHandler(async(req,res) => {
     }
 
     const contact = await Contact.create({
-        name,email,phone
+        name,email,phone,user_id:req.user.id
     });
     res.status(201).json({contact});
 });
 
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async(req,res) => {
     const contact = await Contact.findById(req.params.id);
@@ -51,6 +51,11 @@ const updateContact = asyncHandler(async(req,res) => {
         throw new Error("Contact not found");
     }
 
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("No privileges");
+    }
+    
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id,req.body,{new:true});
 
     res.status(200).json(updatedContact);
@@ -58,7 +63,7 @@ const updateContact = asyncHandler(async(req,res) => {
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async(req,res) => {
 
@@ -68,6 +73,11 @@ const deleteContact = asyncHandler(async(req,res) => {
         throw new Error("Contact not found");
     }
 
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("No privileges");
+    }
+    
     await Contact.findByIdAndDelete(req.params.id);
 
     res.status(200).json(contact);
